\name{coef.strat}
\alias{coef.strat}
\title{Coefficients of a Strategic Model object}
\usage{\method{coef}{strat}(object, ...)
}
\description{Get coefficients from a strategic model}
\details{Extracts the coefficient vector from a fitted model of class \code{strat}.}
\value{numeric vector containing coefficients}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{object}{a fitted model of class \code{strat}}
\item{...}{other arguments, currently ignored}
}

