% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpsmarch2015}
\alias{cpsmarch2015}
\title{A Subset of March CPS 2015 Sample}
\format{A data frame with 14358 rows and 5 variables: \describe{
  \item{income}{personal market income, in US dollars}
  \item{big_class}{big class membership}
  \item{micro_class}{microclass membership}
  \item{education}{educational attainment}
  \item{weight}{sampling weight given by CPS}
  }}
\usage{
cpsmarch2015
}
\description{
A dataset containing income, big class, microclass, and education of 14,358
male respondents from March CPS 2015
}
\keyword{datasets}

