\docType{data}
\name{ganzeboom}
\alias{ganzeboom}
\title{Ganzeboom's ISCO88/ISEI/SIOPS/EGP conversion table.}
\format{A data frame with XXX rows and 4 variables}
\description{
  This table provides the mapping between the ISCO88,
  SIOPS, ISEI and EGP codes.
}
\details{
  \itemize{ \item isco88. ILO ISCO88 \item siops. Treimans
  SIOPS \item isei. Ganzebooms ISEI \item egp. Goldtorp }
}
\references{
  Ganzeboom H. & Treiman D. (1996). Internationally
  Comparable Measures of Occupational Status for the 1988
  International Standard Classification of Occupations.
  \emph{Social Science Research} 25, pp. 201-239.
}
\keyword{dataset}
\keyword{datasets}

