\name{print.strat}
\alias{print.strat}
\title{Print a Strategic Model object}
\usage{\method{print}{strat}(x, ...)
}
\description{The default method for printing a \code{strat} object.}
\details{Prints the call and coefficients of a fitted strategic model.}
\value{\code{x}, invisibly}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{x}{a fitted model of class \code{strat}}
\item{...}{other arguments, currently ignored}
}

