\name{print.summary.strat}
\alias{print.summary.strat}
\title{print summary}
\usage{print.summary.strat(x, ...)
}
\description{Print output from \code{summary.strat}}
\details{Prints the standard regression results table from a fitted strategic model,
along with the log-likelihood, AIC, and number of observations.}
\value{\code{x}, invisibly.}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{x}{an object of class \code{summary.strat}, typically produced by
\code{summary} on a fitted model of class \code{strat}}
\item{...}{other arguments, currently ignored}
}

