\name{summary.strat}
\alias{summary.strat}
\title{Summarize a strategic model object}
\usage{\method{summary}{strat}(object, ...)
}
\description{The default method for summarizing a \code{strat} object.}
\details{Forms the standard regression results table from a fitted strategic model.
Normally used interactively, in conjunction with
\code{\link{print.summary.strat}}.}
\value{an object of class \code{summary.strat}, containing the coefficient
matrix and other information needed for printing}
\seealso{\code{\link{print.summary.strat}}}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{object}{a fitted model of class \code{strat}}
\item{...}{other arguments, currently ignored}
}

