\name{vcov.strat}
\alias{vcov.strat}
\title{Covariance matrix of a Strategic Model object}
\usage{\method{vcov}{strat}(object, ...)
}
\description{Get covariance matrix from a strategic model}
\details{Extracts the covariance matrix from a fitted model of class \code{strat}.}
\value{covariance matrix}
\author{Brenton Kenkel (\email{brenton.kenkel@gmail.com})}
\arguments{\item{object}{a fitted model of class \code{strat}}
\item{...}{other arguments, currently ignored}
}

