\encoding{UTF-8}
\name{ganzeboom}
\docType{data}
\alias{ganzeboom}
\title{Ganzeboom's SIOPS, ISEI, EGP and ISCO88 conversion table}
     \description{
       This table provides the mapping between the ISCO88, SIOPS, ISEI and EGP codes.
     }
     \usage{data(ganzeboom)}
     \format{A dataframe containing 533 rows.}
     \source{Ganzeboom H. & Treiman D. (1996). Appendix A, pages 221-237.}
     \references{
       Ganzeboom H. & Treiman D. (1996). Internationally Comparable Measures of Occupational Status
		for the 1988 International Standard Classification of Occupations. \emph{Social Science 
		Research} 25, pp. 201-239.
     }
     \keyword{datasets}
