% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.srank}
\alias{print.srank}
\title{Print an object of class \code{srank}}
\usage{
\method{print}{srank}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{srank}}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print an object of class \code{srank}
}

